requirements.txt
streamlit
ultralytics
easyocr
numpy
Pillow
opencv-python-headless
torch
torchvision